Align component structure with tx creation components
