+++ i18n/locales/en/common.json
@@ -278,7 +278,6 @@
   "Members": "Members",
   "Message": "Message",
   "Message (optional)": "Message (optional)",
-  "Message signature aborted on device": "Message signature aborted on device",
   "Min": "Min",
   "Minimize": "Minimize",
   "Missed slot": "Missed slot",
@@ -346,7 +345,7 @@
   "Please carefully write down these 12 words and store them in a safe place.": "Please carefully write down these 12 words and store them in a safe place.",
   "Please check the address": "Please check the address",
   "Please check the highlighted word and make sure it’s correct.": "Please check the highlighted word and make sure it’s correct.",
-  "Please confirm the message on your {{model}}": "Please confirm the message on your {{model}}",
+  "Please confirm the message on your {{deviceModel}}": "Please confirm the message on your {{deviceModel}}",
   "Please confirm the transaction on your {{deviceModel}}": "Please confirm the transaction on your {{deviceModel}}",
   "Please enter a valid public key for each member.": "Please enter a valid public key for each member.",
   "Please enter vote amounts for the delegates you wish to vote for": "Please enter vote amounts for the delegates you wish to vote for",
